scripts/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nettest"
dynamic = ["version"]
description = "Comprehensive network testing and diagnostics tool"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Network Testing Tool Contributors" }
]
keywords = [
    "network",
    "diagnostics",
    "speed-test",
    "ping",
    "latency",
    "dns",
    "mtr",
    "traceroute",
    "monitoring",
    "voip",
    "bufferbloat",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: Utilities",
]

dependencies = [
    "rich>=13.0.0",
]

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0",
]
prometheus = [
    "prometheus-client>=0.19.0",
]
http = [
    "requests>=2.31.0",
]
network = [
    "netifaces>=0.11.0",
]
templates = [
    "jinja2>=3.0.0",
]
full = [
    "pyyaml>=6.0",
    "prometheus-client>=0.19.0",
    "requests>=2.31.0",
    "netifaces>=0.11.0",
    "jinja2>=3.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
nettest = "nettest.cli:main"

[project.urls]
Homepage = "https://github.com/manumanohar11/nettest"
Documentation = "https://github.com/manumanohar11/nettest#readme"
Repository = "https://github.com/manumanohar11/nettest.git"
Issues = "https://github.com/manumanohar11/nettest/issues"
Changelog = "https://github.com/manumanohar11/nettest/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "nettest/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/nettest",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["nettest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["nettest"]
branch = true
omit = [
    "nettest/__main__.py",
    "nettest/tui/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["nettest"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "netifaces",
    "prometheus_client",
]
ignore_missing_imports = true
